Introduce SegmentedPostingList to support text index on parts with >4B rows
Replace `PostingList = roaring::Roaring` with `PostingList = SegmentedPostingList`,
a wrapper that splits row IDs into fixed-size segments (default 2^32), each
backed by a `roaring::Roaring` bitmap. All values are stored as absolute UInt64;
internally dispatched to segment-local UInt32. Set operations (`&`, `|`) work
per-segment independently.

For single-segment parts (<=4B rows), the wrapper degenerates to one Roaring
bitmap with near-zero overhead.

Key changes:
- New `SegmentedPostingList` class (header-only)
- `PostingListBuilder::add` accepts `UInt64` (was `UInt32`), `SmallContainer`
  stores 3 x UInt64 (was 6 x UInt32), removed `PostingListWithContext`
- `IPostingListCodec::encode`/`decode` take `roaring::Roaring &` (codec stays
  32-bit, called per-segment)
- Serialization iterates per-segment; `deserialize` takes `segment_row_begin`
  to place blocks into the correct segment
- Reader uses `toUint64Array` and `PaddedPODArray<UInt64>` for index extraction
- Removed >4B row exception in aggregator and merge row limit for text indexes

+++ src/Storages/MergeTree/MergeTreeReaderTextIndex.h
@@ -110,7 +110,7 @@ private:
     size_t current_row = 0;
     size_t current_mark = 0;
 
-    PaddedPODArray<UInt32> indices_buffer;
+    PaddedPODArray<UInt64> indices_buffer;
     roaring::Roaring analyzed_granules;
     roaring::Roaring may_be_true_granules;
